Issue with OGGM multiprocessing fixed (#44)
* Integration tests for PDE solved added

The test suite is starting to become a little bit more complete.

* Fix merge conflicts

* Muninn added as dependency

Now we can run simulations with MB, run by Muninn. New tests have been added covering those usescases. This means that Huginn should be ready to run simulations for most cases. Now it needs to be extensively tested for different regions and finish polishing the remaining issues.

* Changes for compatibility with ODINN.jl

* Sleipnir and Munnin versions updated

* Clean-up and test reference files changes

* Bump version

* Multiprocessing fixed

* Multiprocessing fixed for OGGM

+++ src/setup/config.jl
@@ -41,7 +41,7 @@ function clean()
  end
 
  function enable_multiprocessing(params::Sleipnir.Parameters)
-    procs::Int = params.simulation.workers
+    procs = params.simulation.workers
     if procs > 0 && params.simulation.multiprocessing
         if nprocs() < procs
             @eval begin
